we try to integrate Angular Universal

+++ src/app/services/token-storage.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from "@angular/core";
+import { Inject, Injectable, PLATFORM_ID } from "@angular/core";
+import { isPlatformBrowser } from '@angular/common';
 
  //const TOKEN_KEY = 'auth_token';
 // const USER_KEY = 'auth_user';
@@ -9,27 +10,37 @@ const USER_KEY = 'auth_user';
     providedIn : 'root'
 })
 export class tokenStorageService {
-    constructor () {}
+    constructor (@Inject(PLATFORM_ID) private platformId: any) {}
 
     signOut(): void {
-        window.sessionStorage.clear();
+        if (isPlatformBrowser(this.platformId)) {
+            window.sessionStorage.clear();
+        }
     }
 
     public saveToken (token:string): void {
-        console.log('dans la fct saveTokent ' + token)
-        window.sessionStorage.removeItem(TOKEN_KEY)
-        window.sessionStorage.setItem(TOKEN_KEY, token )
+        //console.log('dans la fct saveTokent ' + token)
+        if (isPlatformBrowser(this.platformId)) {
+            window.sessionStorage.removeItem(TOKEN_KEY)
+            window.sessionStorage.setItem(TOKEN_KEY, token )
+        }
     }
 
     public getToken(): string | null {
         // var tiof = sessionStorage.getItem(TOKEN_KEY);
         // console.log(tiof)
-        return window.sessionStorage.getItem(TOKEN_KEY);
+        let elt!: string|null;
+        if (isPlatformBrowser(this.platformId)) {
+            elt = window.sessionStorage.getItem(TOKEN_KEY);
+          }
+        return elt
     }
 
     public saveUser(user: string): void {
-        window.sessionStorage.removeItem(USER_KEY);
-        window.sessionStorage.setItem(USER_KEY, JSON.stringify(user))
+        if (isPlatformBrowser(this.platformId)) {        
+            window.sessionStorage.removeItem(USER_KEY);
+            window.sessionStorage.setItem(USER_KEY, JSON.stringify(user))
+          }
     } 
 
     public getUser (): any {
